Switch from genereate-99-custom.py to tripleo-standalone-vars
In [1] genereate-99-custom.py has been replaced by
tripleo-standalone-vars. The former output an INI
inventory and the later outputs a YAML inventory.

Update export.py to copy the new script to the
standalone controller, run it on the standalone
controller, and then copy the output back.

Update missing_vars to contain leading whitespace
so it can contanenate with the output of the new
script and still produce valid yaml.

Update unapplied-vars.yml to extract the missing
variables from YAML not INI.

Add a .gitignore to ignore the 99-standalone-vars file
which is generated by tripleo-standalone-vars.

[1] https://review.opendev.org/c/openstack/tripleo-ansible/+/840509