Change the order of operations in crc/glance_dev.sh
This script assumes glance.sh was run first.

It depends on resources created by ~/install_yamls/Makefile
which are created when you run glance.sh. This script then
replaces the default operator with a copy built from what
was checked out of the glance-operator git.

Check out from git before cleaning since cleaning needs to
be completed by the new operator. Echo a message to remind
the user of this.