Use repo copy of core_v1beta1_openstackcontrolplane.yaml
Keep a copy of the meta operator CR in the repo and use it
for scale down. Normally this file is in:

~/install_yamls/out/openstack/openstack/cr/core_v1beta1_openstackcontrolplane.yaml

and is produced by running 'make openstack_deploy'. In the
alternaitve testing method described in

9747a2b86898f24c533422f351bbd04b32be4fb9

'make openstack_deploy' is not run so we can't just scale
down using a file that doesn't yet exist. Thus, we keep a
copy of it in the repo for when using the alternaitve testing
method.