Do not hard code names to be deleted
Instead of hard coding the name, use the output of a
loop which greps on "inheritance" to get the name of
the objects (nodes, inventories, etc) to delete.

The string "inheritance" will be in the names because
the inventory is built using the instance.Name of the
Node and the Node is built using the instance.Name of
the Role and the role to test it all has the string
"inheritance" in it (see role.yaml).